+++ src/tetris/block_tests.hh
@@ -8,10 +8,10 @@
 
 #include "tetris/main.hh"
 
-constexpr usize BLOCK_TEST_BODY_COUNT = available_bodies.size();
-constexpr usize BLOCK_TEST_MAX_BLOCKS = ARBITRARY_FALLING_BODY_BLOCK_SIZE_LIMIT;
-constexpr usize BLOCK_TEST_MAX_VERTICES = BLOCK_TEST_MAX_BLOCKS * gfx::unit_cube_vertices.size;
-constexpr usize BLOCK_TEST_MAX_INDICES = BLOCK_TEST_MAX_BLOCKS * gfx::unit_cube_indices.size;
+constexpr s64 BLOCK_TEST_BODY_COUNT = available_bodies.size();
+constexpr s64 BLOCK_TEST_MAX_BLOCKS = ARBITRARY_FALLING_BODY_BLOCK_SIZE_LIMIT;
+constexpr s64 BLOCK_TEST_MAX_VERTICES = BLOCK_TEST_MAX_BLOCKS * gfx::unit_cube_vertices.size;
+constexpr s64 BLOCK_TEST_MAX_INDICES = BLOCK_TEST_MAX_BLOCKS * gfx::unit_cube_indices.size;
 constexpr f32 BLOCK_TEST_CUBE_SCALE = 0.65f;
 constexpr f32 BLOCK_TEST_CUBE_GAP = 0.1f;
 constexpr f32 BLOCK_TEST_BLOCK_SPACING = 2.f + BLOCK_TEST_CUBE_GAP / BLOCK_TEST_CUBE_SCALE;
@@ -29,7 +29,7 @@ auto make_block_test_meshes() -> Static_Array<gfx::Mesh, BLOCK_TEST_BODY_COUNT>&
     if (block_test_meshes_initialized) return block_test_meshes;
     block_test_meshes_initialized = true;
 
-    usize body_index = 0;
+    s64 body_index = 0;
     for (const auto& body : available_bodies) {
         auto& body_storage = block_test_mesh_storage[body_index];
         s64 vertex_count = 0;
@@ -43,7 +43,7 @@ auto make_block_test_meshes() -> Static_Array<gfx::Mesh, BLOCK_TEST_BODY_COUNT>&
                 cast(f32)layer - cast(f32)center.z,
             };
 
-            const usize vertex_offset = vertex_count;
+            const s64 vertex_offset = vertex_count;
             for (const auto& source_vertex : gfx::DEBUG_CUBE.vertices) {
                 auto vertex = source_vertex;
                 vertex.position += offset * BLOCK_TEST_BLOCK_SPACING;
@@ -85,7 +85,7 @@ auto blocks_tests_main() -> void {
     constexpr f32 MOUSE_ROTATION_SENSITIVITY = 0.01f;
     constexpr auto TARGET_TICKS = ticks_per_frame(144);
 
-    usize body_index = 0;
+    s64 body_index = 0;
     u64 last_tick = get_ticks();
     u64 body_started_at = last_tick;
     gfx::Mesh_Instance body_instance{
